Move text_box code up one level
